+++ src/extension.ts
@@ -3,6 +3,11 @@
 import * as vscode from "vscode";
 import { FilterFormatter } from "./formatter/formatter";
 import { FilterSymbolProvider } from "./outline/filterSymbolProvider";
+import {
+  registerDiagnostics,
+  validateDocument,
+} from "./diagnostics/filterDiagnostics";
+import { FilterCodeActionProvider } from "./diagnostics/filterCodeActions";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -174,6 +179,21 @@ export function activate(context: vscode.ExtensionContext) {
   // if (activeEditor) {
   //   updateDecorations();
   // }
+
+  // Register diagnostics
+  registerDiagnostics(context);
+
+  // Register code actions
+  context.subscriptions.push(
+    vscode.languages.registerCodeActionsProvider(
+      "poe2-filter",
+      new FilterCodeActionProvider(),
+      {
+        providedCodeActionKinds:
+          FilterCodeActionProvider.providedCodeActionKinds,
+      }
+    )
+  );
 }
 
 // This method is called when your extension is deactivated